requirements.txt
# Core dependencies (security updated)
selenium==4.15.0
undetected-chromedriver==3.5.4
webdriver-manager==4.0.1
Pillow==10.3.0
deepface==0.0.79
opencv-python==4.8.1.78
numpy==1.24.3

# Security & Environment
python-dotenv==1.0.0
cryptography==41.0.7
requests==2.31.0
